apps/ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.2
pillow==10.1.0
faiss-cpu==1.7.4
# CLIP - optional, falls back to stub if not available
# open-clip-torch==2.23.0
# torch==2.1.2
ruff==0.1.9
mypy==1.7.1

